+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/OptionComponent.java
@@ -1,7 +1,11 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
+import org.hamcrest.Matchers;
+import org.openqa.selenium.WebElement;
+
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.ClickableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.SeleniumAsserts;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
@@ -14,6 +18,20 @@ public class OptionComponent extends AbstractSeleniumComponent implements Clicka
     public OptionComponent(SeleniumComponent parent, WebElementSelector selector)
     {
         super(parent, selector);
+    }
+
+    public void select()
+    {
+        if (!isSelected())
+        {
+            SeleniumAsserts.assertThatSoon("Select option", () -> {
+                WebElement element = element();
+
+                element.click();
+
+                return element.isSelected();
+            }, Matchers.is(true));
+        }
     }
 
     public boolean isSelected()